Restore -Wsign-compare. TO DO: changes are quick and dirty here. Proepr cleaning would obviously be to not cast to int systematically.

+++ src/RodPair.cc
@@ -337,11 +337,11 @@ void StraightRodPair::buildModules(Container& modules, const RodTemplate& rodTem
   for (int i=0; i<(int)posList.size(); i++, parity = -parity) {
     BarrelModule* mod;
     if (!mezzanine() && (startZMode() == StartZMode::MODULECENTER) && (direction == BuildDir::LEFT)) { // skips the central module information.
-      mod = GeometryFactory::make<BarrelModule>(i < (rodTemplate.size() - 1) ? *rodTemplate[i+1].get() : *rodTemplate.rbegin()->get());
+      mod = GeometryFactory::make<BarrelModule>(i < ((int)rodTemplate.size() - 1) ? *rodTemplate[i+1].get() : *rodTemplate.rbegin()->get());
       mod->myid(i+2);
     }
     else {
-      mod = GeometryFactory::make<BarrelModule>(i < rodTemplate.size() ? *rodTemplate[i].get() : *rodTemplate.rbegin()->get());
+      mod = GeometryFactory::make<BarrelModule>(i < (int)rodTemplate.size() ? *rodTemplate[i].get() : *rodTemplate.rbegin()->get());
       mod->myid(i+1);
     }
     mod->side(side);
@@ -418,7 +418,7 @@ void TiltedRodPair::buildModules(Container& modules, const RodTemplate& rodTempl
   if (tmspecs.empty()) return;
   int i = 0;
   if (direction == BuildDir::LEFT && fabs(tmspecs[0].z) < 0.5) { i = 1; it++; } // this skips the first module if we're going left (i.e. neg rod) and z=0 because it means the pos rod has already got a module there
-  for (; i < tmspecs.size(); i++, ++it) {
+  for (; i < (int)tmspecs.size(); i++, ++it) {
     //std::cout << "i = " << i << std::endl;
     //std::cout << "tmspecs[i].r = " << tmspecs[i].r << std::endl;
     //std::cout << "tmspecs[i].z = " << tmspecs[i].z << std::endl;
